+++ Nice3point.Revit.Extensions/UnitExtensions.cs
@@ -118,4 +118,110 @@ public static class UnitExtensions
         return UnitUtils.ConvertFromInternalUnits(radians, UnitTypeId.Degrees);
 #endif
     }
+#if R19 || R20 || R21
+
+    /// <summary>Formats a number with units into a string</summary>
+    /// <param name="document">Document that stores units</param>
+    /// <param name="unitType">The unit type of the value to format</param>
+    /// <param name="value">The value to format, in Revit's internal units</param>
+    /// <param name="maxAccuracy">
+    ///     True if the value should be rounded to an increased accuracy level appropriate for editing or understanding the precise value stored in the model.
+    ///     False if the accuracy specified by the FormatOptions should be used, appropriate for printed drawings
+    /// </param>
+    /// <param name="forEditing">
+    ///     True if the formatting should be modified as necessary so that the formatted string can be successfully parsed, for example by suppressing digit grouping.
+    ///     False if unmodified settings should be used, suitable for display only
+    /// </param>
+    /// <returns>The formatted string</returns>
+    /// <exception cref="T:Autodesk.Revit.Exceptions.ArgumentException">
+    ///     unitType is an invalid unit type. See UnitUtils.IsValidUnitType() and UnitUtils.GetValidUnitTypes().
+    ///     Or the given value for value is not finite
+    /// </exception>
+    /// <exception cref="T:Autodesk.Revit.Exceptions.ArgumentOutOfRangeException">
+    ///     A value passed for an enumeration argument is not a member of that enumeration
+    /// </exception>
+    [Pure]
+#if R21
+    [Obsolete("This method is deprecated in Revit 2021")]
+#endif
+    public static string FormatUnit(this Document document, UnitType unitType, double value, bool maxAccuracy, bool forEditing)
+    {
+        return UnitFormatUtils.Format(document.GetUnits(), unitType, value, maxAccuracy, forEditing);
+    }
+
+    /// <summary>Formats a number with units into a string</summary>
+    /// <param name="document">Document that stores units</param>
+    /// <param name="unitType">The unit type of the value to format</param>
+    /// <param name="value">The value to format, in Revit's internal units</param>
+    /// <param name="maxAccuracy">
+    ///     True if the value should be rounded to an increased accuracy level appropriate for editing or understanding the precise value stored in the model.
+    ///     False if the accuracy specified by the FormatOptions should be used, appropriate for printed drawings
+    /// </param>
+    /// <param name="forEditing">
+    ///     True if the formatting should be modified as necessary so that the formatted string can be successfully parsed, for example by suppressing digit grouping.
+    ///     False if unmodified settings should be used, suitable for display only
+    /// </param>
+    /// <param name="options">Additional formatting options</param>
+    /// <returns>The formatted string</returns>
+    /// <exception cref="T:Autodesk.Revit.Exceptions.ArgumentException">
+    ///     unitType is an invalid unit type. See UnitUtils.IsValidUnitType() and UnitUtils.GetValidUnitTypes().
+    ///     Or the given value for value is not finite.
+    ///     Or the display unit in the FormatOptions in formatValueOptions is not a valid display unit for unitType.
+    ///     See UnitUtils.IsValidDisplayUnit(UnitType, DisplayUnitType) and UnitUtils.GetValidDisplayUnits(UnitType)
+    /// </exception>
+    /// <exception cref="T:Autodesk.Revit.Exceptions.ArgumentOutOfRangeException">
+    ///     A value passed for an enumeration argument is not a member of that enumeration
+    /// </exception>
+    [Pure]
+#if R21
+    [Obsolete("This method is deprecated in Revit 2021")]
+#endif
+    public static string FormatUnit(this Document document, UnitType unitType, double value, bool maxAccuracy, bool forEditing, FormatValueOptions options)
+    {
+        return UnitFormatUtils.Format(document.GetUnits(), unitType, value, maxAccuracy, forEditing, options);
+    }
+#endif
+#if R21_OR_GREATER
+
+    /// <summary>Formats a number with units into a string</summary>
+    /// <param name="document">Document that stores units</param>
+    /// <param name="specTypeId">Identifier of the spec of the value to format</param>
+    /// <param name="value">The value to format, in Revit's internal units</param>
+    /// <param name="forEditing">
+    ///     True if the formatting should be modified as necessary so that the formatted string can be successfully parsed, for example by suppressing digit grouping.
+    ///     False if unmodified settings should be used, suitable for display only
+    /// </param>
+    /// <returns>The formatted string</returns>
+    /// <exception cref="T:Autodesk.Revit.Exceptions.ArgumentException">
+    ///     specTypeId is not a measurable spec identifier. See UnitUtils.IsMeasurableSpec(ForgeTypeId).
+    ///     Or the given value for value is not finite
+    /// </exception>
+    [Pure]
+    public static string FormatUnit(this Document document, ForgeTypeId specTypeId, double value, bool forEditing)
+    {
+        return UnitFormatUtils.Format(document.GetUnits(), specTypeId, value, forEditing);
+    }
+
+    /// <summary>Formats a number with units into a string</summary>
+    /// <param name="document">Document that stores units</param>
+    /// <param name="specTypeId">Identifier of the spec of the value to format</param>
+    /// <param name="value">The value to format, in Revit's internal units</param>
+    /// <param name="forEditing">
+    ///     True if the formatting should be modified as necessary so that the formatted string can be successfully parsed, for example by suppressing digit grouping.
+    ///     False if unmodified settings should be used, suitable for display only
+    /// </param>
+    /// <param name="options">Additional formatting options</param>
+    /// <returns>The formatted string</returns>
+    /// <exception cref="T:Autodesk.Revit.Exceptions.ArgumentException">
+    ///     specTypeId is not a measurable spec identifier. See UnitUtils.IsMeasurableSpec(ForgeTypeId).
+    ///     Or the given value for value is not finite.
+    ///     Or the unit in the FormatOptions in formatValueOptions is not a valid unit for specTypeId.
+    ///     See UnitUtils.IsValidUnit(ForgeTypeId, ForgeTypeId) and UnitUtils.GetValidUnits(ForgeTypeId)
+    /// </exception>
+    [Pure]
+    public static string FormatUnit(this Document document, ForgeTypeId specTypeId, double value, bool forEditing, FormatValueOptions options)
+    {
+        return UnitFormatUtils.Format(document.GetUnits(), specTypeId, value, forEditing, options);
+    }
+#endif
 }